+++ src/ts/components/settingspanelpageopenbutton.ts
@@ -1,0 +1,21 @@
+import {UIInstanceManager} from '../uimanager';
+import {SettingsPanelPageNavigatorButton, SettingsPanelPageNavigatorConfig} from './settingspanelpagenavigatorbutton';
+
+export class SettingsPanelPageOpenButton extends SettingsPanelPageNavigatorButton {
+  constructor(config: SettingsPanelPageNavigatorConfig) {
+    super(config);
+
+    this.config = this.mergeConfig(config, {
+      cssClass: 'ui-settingspanelpageopenbutton',
+      text: 'open',
+    }, this.config);
+  }
+
+  configure(player: bitmovin.PlayerAPI, uimanager: UIInstanceManager): void {
+    super.configure(player, uimanager);
+
+    this.onClick.subscribe(() => {
+      this.pushTargetPage();
+    });
+  }
+}