+++ src/ts/components/subtitlesettings/subtitlesettingslabel.ts
@@ -1,15 +1,15 @@
 import {LabelConfig} from '../label';
 import {Container, ContainerConfig} from '../container';
-import {SubtitleSettingsOpenButton} from './subtitlesettingsopenbutton';
 import {DOM} from '../../dom';
+import {SettingsPanelPageOpenButton} from '../settingspanelpageopenbutton';
 
 export interface SubtitleSettingsLabelConfig extends LabelConfig {
-  opener: SubtitleSettingsOpenButton;
+  opener: SettingsPanelPageOpenButton;
 }
 
 export class SubtitleSettingsLabel extends Container<ContainerConfig> {
 
-  private opener: SubtitleSettingsOpenButton;
+  private opener: SettingsPanelPageOpenButton;
 
   private text: string;